@@ -27,6 +27,7 @@
     "babel-core": "^6.26.3",
     "browser-sync": "^2.26.10",
     "colors": "^1.4.0",
+    "css-loader": "^6.7.1",
     "del": "^5.1.0",
     "fancy-log": "^1.3.3",
     "gulp": "^4.0.2",
@@ -39,8 +40,10 @@
     "gulp-plumber": "^1.2.1",
     "gulp-size": "^4.0.1",
     "sass-loader": "^13.0.2",
-    "vue": "^2.7.7",
-    "vue-loader": "^15.10.0"
+    "vue": "^3.2.36",
+    "vue-loader": "^17.0.0",
+    "vue-style-loader": "^4.1.3",
+    "vue-template-compiler": "^2.7.7"
   },
   "devDependencies": {
     "@babel/core": "^7.17.9",
